syncer(dm): move location recorder to binlogstream package
